TIMOB-13183 CodeProcessor: Merge require and common-globals plugins into ti-api-provider
