--- plugins/common-globals/README.md
@@ -1,15 +1,0 @@
-Common Globals Plugin
-=====================
-
-## Overview
-
-The Common Globals plugin is a provider plugin and adds support for APIs that aren't part of the JavaScript spec,
-but aren't part of the Titanium spec either, such as ```console.log```.
-
-## Options
-
-No options
-
-## Output
-
-* **name** _string_ Always equals "common-globals"
