--- plugins/common-globals/package.json
@@ -1,13 +1,0 @@
-{
-	"author": "Bryan Hughes <bhughes@appcelerator.com> (https://github.com/bryan-m-hughes/)",
-	"name": "common-globals",
-	"displayName": "Common Globals",
-	"description": "Injects common global properies and functions not in Titanium or the ECMAScript spec, such as console",
-	"titaniumCodeProcessorPlugin": true,
-	"main": "lib/CommonGlobals.js",
-	"dependencies": {},
-	"options": {},
-	"engines": {
-		"node": ">=0.8"
-	}
-}
