--- plugins/require-provider/README.md
@@ -1,15 +1,0 @@
-Require Provider Plugin
-=======================
-
-## Overview
-
-The Require Provider is a provider plugin that provides an implementation of the CommonJS-like mechanism used in
-Titanium Mobile.
-
-## Options
-
-* **platform** _string_ The name of the platform being analyzed. This is used to determine which platform-specific folder to look for files in in the ```Resources``` folder
-
-## Output
-
-* **name** _string_ Always equals "require-provider"
