--- plugins/require-provider/package.json
@@ -1,54 +1,0 @@
-{
-	"author": "Bryan Hughes <bhughes@appcelerator.com> (https://github.com/bryan-m-hughes/)",
-	"name": "require-provider",
-	"displayName": "Require Provider",
-	"description": "Provides a Titanium compliant implementation of require()",
-	"titaniumCodeProcessorPlugin": true,
-	"main": "lib/RequireProvider.js",
-	"dependencies": {},
-	"options": {
-		"platform": {
-			"description": "The platform to use in evaluating require paths",
-			"types": [{
-				"type": "string",
-				"allowedValues": [
-					"android",
-					"mobileweb",
-					"iphone",
-					"ipad",
-					"blackberry"
-				]
-			}],
-			"required": true
-		},
-		"modules": {
-			"description": "The set of modules the project uses",
-			"types": [{
-				"type": "object",
-				"properties": {
-					"<platform name or 'commonjs'>": {
-						"types": [{
-							"type": "object",
-							"properties": {
-								"<module name>": {
-									"types": [{
-										"type": "string",
-										"description": "The path to the module. Only valid for CommonJS modules"
-									},{
-										"type": "null",
-										"description": "Native modules are not evaluated, so the value should be null"
-									}]
-								}
-							}
-						}],
-						"required": false
-					}
-				}
-			}],
-			"required": false
-		}
-	},
-	"engines": {
-		"node": ">=0.8"
-	}
-}
